# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NadlanConnect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      tags: [Listings]
      operationId: listListings
      summary: Search and filter listings
      parameters:
        - name: ville
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: minPrice
          in: query
          required: false
          schema:
            type: integer
        - name: maxPrice
          in: query
          required: false
          schema:
            type: integer
        - name: minSurface
          in: query
          required: false
          schema:
            type: integer
        - name: maxSurface
          in: query
          required: false
          schema:
            type: integer
        - name: nbPieces
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: ownerId
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: List of listings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingsPage'
    post:
      tags: [Listings]
      operationId: createListing
      summary: Create a new listing (agent or developer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingInput'
      responses:
        '201':
          description: Created listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /listings/{listingId}:
    get:
      tags: [Listings]
      operationId: getListing
      summary: Get a single listing with images
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Listing detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingDetail'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [Listings]
      operationId: updateListing
      summary: Update a listing (owner only)
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingUpdate'
      responses:
        '200':
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
    delete:
      tags: [Listings]
      operationId: deleteListing
      summary: Delete a listing (owner or admin)
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteSuccess'
  /listings/{listingId}/images:
    post:
      tags: [Listings]
      operationId: addListingImage
      summary: Add an image to a listing
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingImageInput'
      responses:
        '201':
          description: Image added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingImage'
  /listings/{listingId}/images/{imageId}:
    delete:
      tags: [Listings]
      operationId: deleteListingImage
      summary: Remove an image from a listing
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
        - name: imageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteSuccess'
  /listings/{listingId}/publish:
    post:
      tags: [Listings]
      operationId: publishListing
      summary: Publish a draft listing
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Published listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'

  # ── Favorites ─────────────────────────────────────────────────────────────
...
  /admin/listings:
    get:
      tags: [Dashboard]
      operationId: adminListListings
      summary: Admin view of all listings (any status)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: All listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Listing'
  /admin/listings/{listingId}/status:
    patch:
      tags: [Dashboard]
      operationId: adminUpdateListingStatus
      summary: Admin moderate a listing status
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingStatusUpdate'
      responses:
        '200':
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
components:
...
  schemas:
...
    Listing:
      type: object
      required: [id, ownerId, type, title, ville, surface, nbPieces, price, status, createdAt]
      properties:
        id:
          type: integer
        ownerId:
          type: string
        ownerName:
          type: [string, 'null']
        ownerAvatar:
          type: [string, 'null']
        type:
          type: string
          enum: [resale, new_development]
        title:
          type: string
        description:
          type: [string, 'null']
        ville:
          type: string
        quartier:
          type: [string, 'null']
        surface:
          type: integer
        nbPieces:
          type: integer
        etage:
          type: ['integer', 'null']
        price:
          type: integer
        estimatedPrice:
          type: ['integer', 'null']
        investmentScore:
          type: ['integer', 'null']
        status:
          type: string
          enum: [draft, published, sold, archived]
        coverImageUrl:
          type: ['string', 'null']
        createdAt:
          type: string
          format: date-time
    ListingDetail:
      type: object
      required: [listing, images]
      properties:
        listing:
          $ref: '#/components/schemas/Listing'
        images:
          type: array
          items:
            $ref: '#/components/schemas/ListingImage'
        isFavorited:
          type: boolean
    ListingInput:
      type: object
      required: [type, title, ville, surface, nbPieces, price]
      properties:
        type:
          type: string
          enum: [resale, new_development]
        title:
          type: string
          minLength: 1
        description:
          type: string
        ville:
          type: string
          minLength: 1
        quartier:
          type: string
        surface:
          type: integer
          minimum: 1
        nbPieces:
          type: integer
          minimum: 1
        etage:
          type: integer
        price:
          type: integer
          minimum: 1
    ListingUpdate:
      type: object
      properties:
        type:
          type: string
          enum: [resale, new_development]
        title:
          type: string
        description:
          type: string
        ville:
          type: string
        quartier:
          type: string
        surface:
          type: integer
        nbPieces:
          type: integer
        etage:
          type: integer
        price:
          type: integer
        status:
          type: string
          enum: [draft, published, sold, archived]
    ListingStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [draft, published, sold, archived]
    ListingsPage:
      type: object
      required: [listings, total]
      properties:
        listings:
          type: array
          items:
            $ref: '#/components/schemas/Listing'
        total:
          type: integer
...
    ListingImage:
      type: object
      required: [id, listingId, url, position]
      properties:
        id:
          type: integer
        listingId:
          type: integer
        url:
          type: string
        position:
          type: integer
    ListingImageInput:
      type: object
      required: [url, position]
      properties:
        url:
          type: string
        position:
          type: integer

    # ── Favorites ──
...
    DeleteSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string